refactor(infra): AWS readiness, multi-pasarela, RLS, ISR, rate limiting y CI/CD
Fase 1: Preparacion para AWS
- Split views de websites en paquete modular
- Dockerfile multi-stage para frontend (standalone)
- S3/CloudFront storage condicional en settings
- Gunicorn + dependencias de produccion

Fase 2: Seguridad y escalabilidad
- contextvars async-safe para tenant context
- JSON structured logging (produccion)
- Row-Level Security (RLS) en PostgreSQL con cast UUID
- Rate limiting por tenant y usuario (TenantRateThrottle)

Fase 3: Pagos multi-pais y pricing
- Modelo PaymentGateway multi-pasarela por tenant
- tax_rate y tax_name configurables por tenant
- Planes de pricing en COP (Starter/Pro/Business)
- ISR + generateMetadata para SEO dinamico
- Pipeline CI/CD para AWS ECS/Fargate
