+++ .env.example
@@ -1,5 +1,20 @@
 # Variables para Docker Compose (desarrollo local)
 # Copiar a .env y ajustar valores
+
+# --- Base de datos ---
 POSTGRES_DB=nerbis_db
 POSTGRES_USER=admin
 POSTGRES_PASSWORD=cambiar_esta_contraseña
+
+# --- Tenant por defecto (desarrollo) ---
+# Slug del tenant por defecto para desarrollo local; cámbialo al tenant correspondiente
+DEFAULT_TENANT_SLUG=demo
+
+# --- AWS S3 (solo produccion) ---
+# AWS_STORAGE_BUCKET_NAME=nerbis-media
+# AWS_S3_REGION_NAME=us-east-1
+# AWS_CLOUDFRONT_DOMAIN=d1234567.cloudfront.net
+
+# --- Frontend (solo produccion) ---
+# NEXT_PUBLIC_API_HOSTNAME=api.nerbis.com
+# NEXT_PUBLIC_CDN_HOSTNAME=d1234567.cloudfront.net
