+++ backend/Dockerfile
@@ -26,5 +26,5 @@ COPY . .
 # Exponer puerto
 EXPOSE 8000
 
-# Comando por defecto
-CMD ["python", "manage.py", "runserver", "0.0.0.0:8000"]
+# Comando por defecto (Gunicorn para producción)
+CMD ["gunicorn", "config.wsgi:application", "--bind", "0.0.0.0:8000", "--workers", "4", "--threads", "2", "--timeout", "120"]
