+++ backend/core/throttles.py
@@ -1,11 +1,69 @@
 # backend/core/throttles.py
-# Rate limiting para endpoints de autenticación.
-# Previene brute force, spam de OTP y enumeración de cuentas.
+"""
+Rate limiting para endpoints de autenticación y API general.
+Previene brute force, spam de OTP, enumeración de cuentas,
+y abuso del API por tenant.
+"""
 
 from django.core.cache import caches
 from rest_framework.throttling import SimpleRateThrottle
 
 throttle_cache = caches["throttle"]
+
+# Límites de API por plan del tenant (requests/hora)
+PLAN_RATE_LIMITS = {
+    "trial": "500/hour",
+    "basic": "1000/hour",
+    "professional": "5000/hour",
+    "enterprise": "20000/hour",
+}
+
+
+class TenantRateThrottle(SimpleRateThrottle):
+    """
+    Rate limit por tenant — evita que un tenant consuma todos los recursos.
+    El límite varía según el plan contratado.
+    """
+
+    scope = "tenant"
+    cache = throttle_cache
+
+    def get_cache_key(self, request, view):
+        tenant = getattr(request, "tenant", None)
+        if not tenant:
+            return None
+        return f"throttle_tenant_{tenant.id}"
+
+    def get_rate(self):
+        """Obtener rate según el plan del tenant."""
+        if not hasattr(self, "_request"):
+            return PLAN_RATE_LIMITS["trial"]
+        tenant = getattr(self._request, "tenant", None)
+        if not tenant:
+            return PLAN_RATE_LIMITS["trial"]
+        return PLAN_RATE_LIMITS.get(tenant.plan, PLAN_RATE_LIMITS["basic"])
+
+    def allow_request(self, request, view):
+        self._request = request
+        return super().allow_request(request, view)
+
+
+class UserRateThrottle(SimpleRateThrottle):
+    """
+    Rate limit por usuario autenticado — 120 requests/min.
+    Previene que un solo usuario abuse del API.
+    """
+
+    scope = "user"
+    cache = throttle_cache
+
+    def get_cache_key(self, request, view):
+        if request.user and request.user.is_authenticated:
+            return f"throttle_user_{request.user.pk}"
+        return self.cache_format % {
+            "scope": self.scope,
+            "ident": self.get_ident(request),
+        }
 
 
 class LoginThrottle(SimpleRateThrottle):
