+++ backend/requirements.txt
@@ -60,3 +60,19 @@ webauthn>=2.5.0
 # ===================================
 pyotp>=2.9.0
 qrcode[pil]>=7.4.2
+
+# ===================================
+# SERVIDOR WSGI (producción)
+# ===================================
+gunicorn>=22.0.0
+
+# ===================================
+# ALMACENAMIENTO EN LA NUBE (S3)
+# ===================================
+django-storages[s3]>=1.14.0
+boto3>=1.35.0
+
+# ===================================
+# LOGGING ESTRUCTURADO (producción)
+# ===================================
+python-json-logger>=2.0.0
