+++ backend/websites/views/preview.py
@@ -1,980 +1,17 @@
-# backend/websites/views.py
-"""
-API Views para el Website Builder.
-
-Endpoints para:
-- Listar y obtener templates
-- Proceso de onboarding
-- Generación de contenido con IA
-- Chat para ediciones
-- Publicación del sitio
-"""
+# backend/websites/views/preview.py
+"""Views para preview y rendering HTML del sitio web."""
 
 import logging
 
 from django.conf import settings as django_settings
-from django.db import models, transaction
-from django.shortcuts import get_object_or_404
-from django.utils import timezone
-from rest_framework import status, viewsets
-from rest_framework.decorators import action
+from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from .models import (
-    ChatMessage,
-    OnboardingQuestion,
-    OnboardingResponse,
-    WebsiteConfig,
-    WebsiteTemplate,
-)
-from .serializers import (
-    BulkOnboardingResponseSerializer,
-    ChatRequestSerializer,
-    GenerateContentRequestSerializer,
-    OnboardingQuestionSerializer,
-    PublishWebsiteSerializer,
-    WebsiteConfigCreateSerializer,
-    WebsiteConfigSerializer,
-    WebsiteTemplateDetailSerializer,
-    WebsiteTemplateListSerializer,
-)
-from .services import AIService
+from ..models import OnboardingResponse, WebsiteConfig
 
 logger = logging.getLogger(__name__)
-
-
-# ===================================
-# TEMPLATES
-# ===================================
-
-
-class WebsiteTemplateViewSet(viewsets.ReadOnlyModelViewSet):
-    """
-    ViewSet para listar y obtener templates de sitios web.
-
-    GET /api/websites/templates/ - Lista todos los templates activos
-    GET /api/websites/templates/{id}/ - Detalle de un template con sus preguntas
-    """
-
-    permission_classes = [IsAuthenticated]
-    queryset = WebsiteTemplate.objects.filter(is_active=True)
-
-    def get_serializer_class(self):
-        if self.action == "retrieve":
-            return WebsiteTemplateDetailSerializer
-        return WebsiteTemplateListSerializer
-
-    def get_queryset(self):
-        queryset = super().get_queryset()
-
-        # Filtrar por industria si se especifica
-        industry = self.request.query_params.get("industry")
-        if industry:
-            queryset = queryset.filter(industry=industry)
-
-        # Filtrar premium si se especifica
-        premium = self.request.query_params.get("premium")
-        if premium is not None:
-            queryset = queryset.filter(is_premium=premium.lower() == "true")
-
-        return queryset.order_by("sort_order", "name")
-
-
-# ===================================
-# WEBSITE CONFIG (Sitio del usuario)
-# ===================================
-
-
-class WebsiteConfigViewSet(viewsets.ModelViewSet):
-    """
-    ViewSet para gestionar la configuración del sitio web del tenant.
-
-    El tenant solo puede tener UN sitio web (OneToOne con Tenant).
-    """
-
-    permission_classes = [IsAuthenticated]
-    serializer_class = WebsiteConfigSerializer
-
-    def get_queryset(self):
-        """Filtra por el tenant del usuario."""
-        user = self.request.user
-        if user.is_superuser:
-            return WebsiteConfig.objects.all()
-        if hasattr(user, "tenant") and user.tenant:
-            return WebsiteConfig.objects.filter(tenant=user.tenant)
-        return WebsiteConfig.objects.none()
-
-    def get_object(self):
-        """Obtiene el sitio del tenant actual (o por ID si es superuser)."""
-        user = self.request.user
-
-        # Si se pasa un ID y es superuser, usar ese ID
-        if "pk" in self.kwargs and user.is_superuser:
-            return get_object_or_404(WebsiteConfig, pk=self.kwargs["pk"])
-
-        # Para usuarios normales, obtener su sitio
-        if hasattr(user, "tenant") and user.tenant:
-            return get_object_or_404(WebsiteConfig, tenant=user.tenant)
-
-        return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_404_NOT_FOUND)
-
-    def perform_update(self, serializer):
-        """Keep onboarding logo_upload in sync when logo changes from editor."""
-        old_logo = (serializer.instance.media_data or {}).get("logo_url", "")
-        instance = serializer.save()
-        new_logo = (instance.media_data or {}).get("logo_url", "")
-        # Sync when logo changed (including removal)
-        if new_logo != old_logo:
-            OnboardingResponse.objects.filter(
-                website_config=instance,
-                question__question_key="logo_upload",
-            ).update(response_value=new_logo or "")
-
-    @action(detail=False, methods=["get"])
-    def my_site(self, request):
-        """
-        GET /api/websites/config/my_site/
-
-        Obtiene el sitio web del tenant actual.
-        Si no existe, retorna información para crearlo.
-        """
-        user = request.user
-        if not hasattr(user, "tenant") or not user.tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=user.tenant)
-            return Response(WebsiteConfigSerializer(config).data)
-        except WebsiteConfig.DoesNotExist:
-            return Response(
-                {
-                    "exists": False,
-                    "message": "No tienes un sitio web. Selecciona un template para comenzar.",
-                    "templates_url": "/api/websites/templates/",
-                }
-            )
-
-
-# ===================================
-# ONBOARDING
-# ===================================
-
-
-class OnboardingView(APIView):
-    """
-    Base class para views de onboarding.
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def _get_tenant(self, request):
-        """Obtiene el tenant del usuario."""
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-class StartOnboardingView(OnboardingView):
-    """
-    POST /api/websites/onboarding/start/
-
-    Inicia el proceso de onboarding seleccionando un template.
-    Crea o actualiza el WebsiteConfig del tenant.
-    """
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        serializer = WebsiteConfigCreateSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        template_id = serializer.validated_data["template_id"]
-        template = WebsiteTemplate.objects.get(id=template_id)
-
-        # Crear o actualizar WebsiteConfig
-        config, created = WebsiteConfig.objects.update_or_create(
-            tenant=tenant,
-            defaults={
-                "template": template,
-                "status": "onboarding",
-                "subdomain": tenant.slug,  # Usar slug del tenant como subdominio inicial
-            },
-        )
-
-        # Obtener preguntas del template
-        questions = self._get_questions_for_template(template)
-
-        return Response(
-            {
-                "config_id": config.id,
-                "template": WebsiteTemplateListSerializer(template).data,
-                "questions": OnboardingQuestionSerializer(questions, many=True).data,
-                "status": config.status,
-                "created": created,
-            },
-            status=status.HTTP_201_CREATED if created else status.HTTP_200_OK,
-        )
-
-    def _get_questions_for_template(self, template):
-        """Obtiene preguntas genéricas + específicas del template."""
-        generic = OnboardingQuestion.objects.filter(template__isnull=True, is_active=True)
-        specific = template.questions.filter(is_active=True)
-
-        # Combinar y ordenar
-        all_questions = list(generic) + list(specific)
-        all_questions.sort(key=lambda q: (q.section, q.sort_order))
-        return all_questions
-
-
-class SaveOnboardingResponsesView(OnboardingView):
-    """
-    POST /api/websites/onboarding/save/
-
-    Guarda las respuestas del onboarding.
-    Acepta guardar parcialmente (por sección) o todas a la vez.
-    """
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        # Verificar que existe un sitio en onboarding
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "Primero debes seleccionar un template"}, status=status.HTTP_400_BAD_REQUEST)
-
-        serializer = BulkOnboardingResponseSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        responses_data = serializer.validated_data["responses"]
-
-        # Guardar respuestas
-        saved_count = 0
-        with transaction.atomic():
-            for question_key, response_value in responses_data.items():
-                question = OnboardingQuestion.objects.get(question_key=question_key, is_active=True)
-
-                OnboardingResponse.objects.update_or_create(
-                    website_config=config, question=question, defaults={"response_value": response_value}
-                )
-                saved_count += 1
-
-                # Sync logo changes to media_data immediately
-                if question_key == "logo_upload" and response_value:
-                    media = dict(config.media_data or {})
-                    media["logo_url"] = response_value
-                    config.media_data = media
-                    config.save(update_fields=["media_data"])
-
-        return Response(
-            {"saved": saved_count, "message": f"Se guardaron {saved_count} respuestas", "config_status": config.status}
-        )
-
-
-class OnboardingStatusView(OnboardingView):
-    """
-    GET /api/websites/onboarding/status/
-
-    Obtiene el estado actual del onboarding.
-    """
-
-    def get(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"status": "not_started", "message": "No has iniciado el proceso. Selecciona un template."})
-
-        # Obtener respuestas guardadas
-        responses = OnboardingResponse.objects.filter(website_config=config).select_related("question")
-
-        responses_dict = {r.question.question_key: r.response_value for r in responses}
-
-        # Obtener preguntas requeridas
-        required_questions = (
-            OnboardingQuestion.objects.filter(is_required=True, is_active=True)
-            .filter(models.Q(template__isnull=True) | models.Q(template=config.template))
-            .values_list("question_key", flat=True)
-        )
-
-        # Verificar completitud
-        answered_required = [key for key in required_questions if key in responses_dict and responses_dict[key]]
-
-        return Response(
-            {
-                "status": config.status,
-                "template": WebsiteTemplateListSerializer(config.template).data,
-                "responses": responses_dict,
-                "progress": {
-                    "total_required": len(required_questions),
-                    "answered_required": len(answered_required),
-                    "is_complete": len(answered_required) == len(required_questions),
-                },
-            }
-        )
-
-
-# ===================================
-# QUICK-START — FLUJO <60s (Fase 2)
-# ===================================
-
-
-class QuickStartView(APIView):
-    """
-    POST /api/websites/onboarding/quick-start/
-
-    Flujo acelerado post-registro: recibe 3 campos y despacha la generacion
-    del sitio completo a una tarea de Celery.
-
-    Combina: auto-resolucion de template + guardar respuestas minimas +
-    despacho asincrono de generacion con IA.
-
-    Campos:
-    - business_description (obligatorio): que hace el negocio + que lo diferencia.
-    - main_services (obligatorio): servicios principales, uno por linea.
-    - business_whatsapp (opcional): se pre-llena con phone del tenant.
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        from celery.result import AsyncResult
-
-        from .tasks import generate_website_content
-
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        from .serializers import QuickStartSerializer
-
-        serializer = QuickStartSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        # 1. Resolver template por industria
-        from core.industry_defaults import get_default_template_slug
-
-        slug = get_default_template_slug(tenant.industry)
-        if not slug:
-            return Response(
-                {"error": "Tu industria aún no tiene un template asignado. Usa el flujo completo."},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        template = WebsiteTemplate.objects.filter(slug=slug, is_active=True).first()
-        if not template:
-            return Response(
-                {"error": f"Template '{slug}' no encontrado. Contacta soporte."},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        # 2. Crear o actualizar WebsiteConfig
-        config, _created = WebsiteConfig.objects.update_or_create(
-            tenant=tenant,
-            defaults={
-                "template": template,
-                "status": "generating",
-                "subdomain": tenant.slug,
-            },
-        )
-
-        # Guard: no permitir generacion duplicada si ya hay una tarea activa
-        if config.generation_task_id:
-            result = AsyncResult(config.generation_task_id)
-            if result.state not in ("SUCCESS", "FAILURE", "REVOKED"):
-                return Response(
-                    {
-                        "error": "Ya hay una generación en progreso",
-                        "task_id": config.generation_task_id,
-                        "status": "generating",
-                    },
-                    status=status.HTTP_409_CONFLICT,
-                )
-
-        # 3. Armar onboarding_responses a partir de los 3 campos + tenant data
-        data = serializer.validated_data
-        responses_dict = {
-            "business_name": tenant.name,
-            "business_description": data["business_description"],
-            "main_services": data["main_services"],
-            "business_whatsapp": data.get("business_whatsapp") or tenant.phone or "",
-            "business_phone": tenant.phone or "",
-            "business_email": tenant.email or "",
-            "business_address": tenant.address or "",
-        }
-        if data.get("website_sections"):
-            responses_dict["website_sections"] = data["website_sections"]
-
-        # 4. Verificar limite de generaciones
-        ai_service = AIService(tenant=tenant, website_config=config)
-        can_generate, used, limit = ai_service.check_usage_limit(tenant)
-        if not can_generate:
-            config.status = "onboarding"
-            config.save(update_fields=["status"])
-            return Response(
-                {
-                    "error": "Has alcanzado el límite de generaciones este mes",
-                    "used": used,
-                    "limit": limit,
-                },
-                status=status.HTTP_402_PAYMENT_REQUIRED,
-            )
-
-        # 5. Despachar tarea asincrona
-        task = generate_website_content.delay(
-            website_config_id=config.id,
-            tenant_id=tenant.id,
-            user_id=request.user.id,
-            onboarding_responses=responses_dict,
-            is_quick_start=True,
-        )
-
-        # Guardar task_id y confirmar estado
-        config.generation_task_id = task.id
-        config.save(update_fields=["generation_task_id"])
-
-        return Response(
-            {
-                "task_id": task.id,
-                "status": "generating",
-                "template": {
-                    "slug": template.slug,
-                    "name": template.name,
-                },
-            },
-            status=status.HTTP_202_ACCEPTED,
-        )
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# GENERACIÓN DE CONTENIDO
-# ===================================
-
-
-class GenerateContentView(APIView):
-    """
-    POST /api/websites/generate/
-
-    Despacha la generacion de contenido del sitio web a una tarea de Celery.
-    Retorna 202 Accepted con el task_id para que el frontend haga polling.
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        from celery.result import AsyncResult
-
-        from .tasks import generate_website_content
-
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "Primero debes completar el onboarding"}, status=status.HTTP_400_BAD_REQUEST)
-
-        serializer = GenerateContentRequestSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        # Guard: no permitir generacion duplicada si ya hay una tarea activa
-        if config.generation_task_id:
-            result = AsyncResult(config.generation_task_id)
-            if result.state not in ("SUCCESS", "FAILURE", "REVOKED"):
-                return Response(
-                    {
-                        "error": "Ya hay una generación en progreso",
-                        "task_id": config.generation_task_id,
-                        "status": "generating",
-                    },
-                    status=status.HTTP_409_CONFLICT,
-                )
-
-        # Verificar limite de uso
-        ai_service = AIService(tenant=tenant, website_config=config)
-        can_generate, used, limit = ai_service.check_usage_limit(tenant)
-
-        if not can_generate:
-            return Response(
-                {
-                    "error": "Has alcanzado el límite de generaciones este mes",
-                    "used": used,
-                    "limit": limit,
-                    "upgrade_url": "/planes/",
-                },
-                status=status.HTTP_402_PAYMENT_REQUIRED,
-            )
-
-        # Obtener respuestas del onboarding
-        responses = OnboardingResponse.objects.filter(website_config=config).select_related("question")
-        responses_dict = {r.question.question_key: r.response_value for r in responses}
-
-        regenerate_section = serializer.validated_data.get("regenerate_section")
-        additional_instructions = serializer.validated_data.get("additional_instructions", "")
-
-        # Despachar tarea asincrona
-        task = generate_website_content.delay(
-            website_config_id=config.id,
-            tenant_id=tenant.id,
-            user_id=request.user.id,
-            onboarding_responses=responses_dict,
-            additional_instructions=additional_instructions,
-            regenerate_section=regenerate_section,
-            is_quick_start=False,
-        )
-
-        # Guardar task_id y actualizar estado
-        config.generation_task_id = task.id
-        config.status = "generating"
-        config.save(update_fields=["status", "generation_task_id"])
-
-        return Response(
-            {
-                "task_id": task.id,
-                "status": "generating",
-            },
-            status=status.HTTP_202_ACCEPTED,
-        )
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-class GenerationStatusView(APIView):
-    """
-    GET /api/websites/generation-status/
-
-    Retorna el estado actual de la generacion de contenido.
-    El frontend hace polling a este endpoint mientras status == "generating".
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_404_NOT_FOUND)
-
-        response_data = {
-            "status": config.status,
-            "task_id": config.generation_task_id,
-        }
-
-        # Incluir datos generados solo cuando la generacion termino exitosamente
-        if config.status == "review":
-            response_data["content_data"] = config.content_data
-            response_data["seo_data"] = config.seo_data
-            response_data["theme_data"] = config.theme_data
-
-        return Response(response_data)
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# CHAT CON IA
-# ===================================
-
-
-class ChatView(APIView):
-    """
-    POST /api/websites/chat/ - Envía mensaje al chat
-    GET /api/websites/chat/ - Obtiene historial del chat
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        """Envía un mensaje al chat para editar el contenido."""
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_400_BAD_REQUEST)
-
-        if config.status not in ["review", "published"]:
-            return Response({"error": "Primero debes generar el contenido inicial"}, status=status.HTTP_400_BAD_REQUEST)
-
-        serializer = ChatRequestSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        message = serializer.validated_data["message"]
-        section_id = serializer.validated_data.get("section_id")
-
-        # Verificar límite de uso
-        ai_service = AIService(tenant=tenant, website_config=config)
-        can_generate, used, limit = ai_service.check_usage_limit(tenant)
-
-        if not can_generate:
-            return Response(
-                {"error": "Has alcanzado el límite de generaciones este mes", "used": used, "limit": limit},
-                status=status.HTTP_402_PAYMENT_REQUIRED,
-            )
-
-        # Obtener historial del chat
-        chat_history = list(
-            ChatMessage.objects.filter(website_config=config).order_by("-created_at")[:20].values("role", "content")
-        )
-        chat_history.reverse()
-
-        # Guardar mensaje del usuario
-        ChatMessage.objects.create(website_config=config, role="user", content=message, section_id=section_id or "")
-
-        try:
-            # Procesar con IA
-            response_message, updated_content, affected_section, tokens_in, tokens_out = ai_service.chat_edit(
-                message=message, current_content=config.content_data, chat_history=chat_history, section_id=section_id
-            )
-
-            # Guardar mensaje del asistente
-            ChatMessage.objects.create(
-                website_config=config,
-                role="assistant",
-                content=response_message,
-                section_id=affected_section or "",
-                changes_made={"updated_content": updated_content} if updated_content else {},
-                tokens_used=tokens_in + tokens_out,
-            )
-
-            # Actualizar contenido si hay cambios
-            if updated_content and affected_section:
-                config.content_data[affected_section] = updated_content
-                config.save(update_fields=["content_data"])
-
-            # Registrar uso
-            ai_service.log_generation(
-                generation_type="edit_content",
-                tokens_input=tokens_in,
-                tokens_output=tokens_out,
-                section_id=affected_section or "",
-                is_successful=True,
-            )
-
-            _, new_used, new_limit = ai_service.check_usage_limit(tenant)
-
-            return Response(
-                {
-                    "message": response_message,
-                    "updated_content": updated_content,
-                    "section_id": affected_section,
-                    "tokens_used": tokens_in + tokens_out,
-                    "remaining_generations": max(0, new_limit - new_used),
-                }
-            )
-
-        except Exception as e:
-            logger.error(f"Error en chat: {e}")
-            return Response(
-                {"error": "Error procesando mensaje. Intenta de nuevo."}, status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-
-    def get(self, request):
-        """Obtiene el historial del chat."""
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"messages": []})
-
-        messages = (
-            ChatMessage.objects.filter(website_config=config)
-            .order_by("created_at")
-            .values("id", "role", "content", "section_id", "created_at")
-        )
-
-        return Response({"messages": list(messages)})
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# SEO AI SUGGESTIONS
-# ===================================
-
-
-class SuggestSeoView(APIView):
-    """
-    POST /api/websites/suggest-seo/
-
-    Genera sugerencias de título y descripción SEO usando IA,
-    basándose en las keywords que el usuario ingresó.
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_400_BAD_REQUEST)
-
-        keywords = request.data.get("keywords", [])
-        business_name = request.data.get("business_name", "")
-        current_title = request.data.get("current_title", "")
-        current_description = request.data.get("current_description", "")
-
-        if not keywords and not business_name:
-            return Response(
-                {"error": "Agrega al menos una palabra clave o el nombre de tu negocio"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        # Check usage limit
-        ai_service = AIService(tenant=tenant, website_config=config)
-        can_generate, used, limit = ai_service.check_usage_limit(tenant)
-
-        if not can_generate:
-            return Response(
-                {
-                    "error": "Has alcanzado el límite de generaciones con IA este mes",
-                    "used": used,
-                    "limit": limit,
-                },
-                status=status.HTTP_402_PAYMENT_REQUIRED,
-            )
-
-        # Generate suggestions
-        suggestions, tokens_in, tokens_out = ai_service.suggest_seo(
-            keywords=keywords,
-            business_name=business_name,
-            current_title=current_title,
-            current_description=current_description,
-        )
-
-        # Log generation
-        ai_service.log_generation(
-            generation_type="seo_suggest",
-            tokens_input=tokens_in,
-            tokens_output=tokens_out,
-            is_successful=True,
-        )
-
-        return Response(
-            {
-                "title": suggestions.get("title", ""),
-                "description": suggestions.get("description", ""),
-                "extra_keywords": suggestions.get("extra_keywords", []),
-            }
-        )
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# UPLOAD DE MEDIA
-# ===================================
-
-
-class UploadWebsiteMediaView(APIView):
-    """
-    POST /api/websites/upload-media/
-
-    Sube una imagen para el website builder (OG image, favicon, etc).
-    Devuelve la URL absoluta del archivo subido.
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    MAX_FILE_SIZE = 5 * 1024 * 1024  # 5MB
-    ALLOWED_IMAGE_TYPES = {
-        "image/jpeg",
-        "image/png",
-        "image/webp",
-        "image/gif",
-        "image/svg+xml",
-        "image/x-icon",
-        "image/vnd.microsoft.icon",
-    }
-
-    ALLOWED_EXTENSIONS = {".jpg", ".jpeg", ".png", ".webp", ".gif", ".svg", ".ico"}
-    # Magic bytes for image format detection
-    _MAGIC_BYTES = {
-        b"\xff\xd8\xff": ".jpg",
-        b"\x89PNG": ".png",
-        b"RIFF": ".webp",  # WebP starts with RIFF....WEBP
-        b"GIF8": ".gif",
-        b"\x00\x00\x01\x00": ".ico",
-        b"\x00\x00\x02\x00": ".ico",
-    }
-
-    def _detect_image_type(self, uploaded_file):
-        """Detect image type from file content (magic bytes), not client headers."""
-        uploaded_file.seek(0)
-        header = uploaded_file.read(12)
-        uploaded_file.seek(0)
-
-        # Check SVG (text-based)
-        if header.lstrip(b"\xef\xbb\xbf").lstrip()[:5] in (b"<?xml", b"<svg "):
-            return ".svg"
-        # WebP: RIFF....WEBP
-        if header[:4] == b"RIFF" and header[8:12] == b"WEBP":
-            return ".webp"
-        for magic, ext in self._MAGIC_BYTES.items():
-            if magic != b"RIFF" and header[: len(magic)] == magic:
-                return ext
-        return None
-
-    def post(self, request):
-        import uuid
-
-        from django.core.files.storage import default_storage
-
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        uploaded_file = request.FILES.get("file")
-        if not uploaded_file:
-            return Response({"error": "No se envió ningún archivo"}, status=status.HTTP_400_BAD_REQUEST)
-
-        # Validar tamaño
-        if uploaded_file.size > self.MAX_FILE_SIZE:
-            return Response({"error": "La imagen es muy grande. El máximo es 5MB."}, status=status.HTTP_400_BAD_REQUEST)
-
-        # Validar tipo real del archivo (magic bytes, no content_type del cliente)
-        detected_ext = self._detect_image_type(uploaded_file)
-        if not detected_ext or detected_ext not in self.ALLOWED_EXTENSIONS:
-            return Response(
-                {"error": "Formato no soportado. Usa JPG, PNG, WebP, GIF o SVG."}, status=status.HTTP_400_BAD_REQUEST
-            )
-
-        purpose = request.data.get("purpose", "general")
-        if purpose not in ("og_image", "favicon", "general"):
-            purpose = "general"
-
-        # Generar nombre seguro con extensión detectada (no la del cliente)
-        ext = detected_ext
-        filename = f"{purpose}_{uuid.uuid4().hex[:8]}{ext}"
-        path = f"websites/{tenant.slug}/{filename}"
-
-        # Guardar archivo
-        saved_path = default_storage.save(path, uploaded_file)
-        file_url = request.build_absolute_uri(f"/media/{saved_path}")
-
-        return Response(
-            {
-                "url": file_url,
-                "path": saved_path,
-            },
-            status=status.HTTP_201_CREATED,
-        )
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# PUBLICACIÓN
-# ===================================
-
-
-class PublishWebsiteView(APIView):
-    """
-    POST /api/websites/publish/
-
-    Publica el sitio web.
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_400_BAD_REQUEST)
-
-        if config.status not in ["review", "published"]:
-            return Response(
-                {"error": "El sitio debe estar en revisión para publicar"}, status=status.HTTP_400_BAD_REQUEST
-            )
-
-        serializer = PublishWebsiteSerializer(data=request.data, context={"tenant": tenant})
-        serializer.is_valid(raise_exception=True)
-
-        # Actualizar subdominio si se proporciona
-        new_subdomain = serializer.validated_data.get("subdomain")
-        if new_subdomain:
-            config.subdomain = new_subdomain
-
-        # Crear snapshot de datos publicados
-        config.published_data = {
-            "content_data": config.content_data or {},
-            "theme_data": config.theme_data or {},
-            "pages_data": config.pages_data or {},
-            "seo_data": config.seo_data or {},
-            "media_data": config.media_data or {},
-        }
-
-        # Publicar
-        config.status = "published"
-        config.published_at = timezone.now()
-        config.save(update_fields=["published_data", "status", "published_at", "subdomain"])
-
-        return Response(
-            {
-                "message": "Sitio publicado exitosamente",
-                "public_url": config.public_url,
-                "status": config.status,
-                "published_at": config.published_at,
-            }
-        )
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# PREVIEW
-# ===================================
 
 
 class PreviewWebsiteView(APIView):
@@ -1019,11 +56,6 @@ class PreviewWebsiteView(APIView):
         if not hasattr(request.user, "tenant") or not request.user.tenant:
             return None
         return request.user.tenant
-
-
-# ===================================
-# PREVIEW RENDER (HTML para iframe)
-# ===================================
 
 
 class PreviewRenderView(APIView):
@@ -5039,363 +4071,6 @@ class PreviewRenderView(APIView):
     </div>
 </section>
 """
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# SECTION MANAGEMENT
-# ===================================
-
-
-class ReorderSectionsView(APIView):
-    """
-    POST /api/websites/sections/reorder/
-
-    Reordena las secciones del content_data.
-    Body: {"order": ["hero", "services", "about", "contact"]}
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_404_NOT_FOUND)
-
-        new_order = request.data.get("order", [])
-        if not new_order or not isinstance(new_order, list):
-            return Response(
-                {"error": "Se requiere 'order' como lista de IDs de sección"}, status=status.HTTP_400_BAD_REQUEST
-            )
-
-        content = config.content_data or {}
-        content["_section_order"] = new_order
-        config.content_data = content
-        config.save(update_fields=["content_data", "updated_at"])
-
-        return Response({"message": "Secciones reordenadas", "order": new_order})
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-class AddSectionView(APIView):
-    """
-    POST /api/websites/sections/add/
-
-    Agrega una sección vacía al content_data.
-    Body: {"section_id": "gallery"}
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_404_NOT_FOUND)
-
-        section_id = request.data.get("section_id", "")
-        if not section_id:
-            return Response({"error": "Se requiere 'section_id'"}, status=status.HTTP_400_BAD_REQUEST)
-
-        # Verificar que la sección exista en el template
-        structure = config.template.structure_schema or {}
-        available = {s["id"] for s in structure.get("sections", [])}
-        if section_id not in available:
-            return Response(
-                {"error": f"La sección '{section_id}' no existe en este template"}, status=status.HTTP_400_BAD_REQUEST
-            )
-
-        content = config.content_data or {}
-        if section_id in content:
-            return Response({"error": f"La sección '{section_id}' ya existe"}, status=status.HTTP_400_BAD_REQUEST)
-
-        # Usar contenido inicial del frontend si viene, o defaults
-        initial_content = request.data.get("initial_content")
-        variant = request.data.get("variant")
-
-        if initial_content and isinstance(initial_content, dict):
-            content[section_id] = initial_content
-        else:
-            section_defaults = {
-                "header": {"logo_text": "", "cta_text": "Contáctanos", "cta_link": "#contact"},
-                "hero": {"title": "", "subtitle": "", "cta_text": "", "cta_link": "#"},
-                "about": {"title": "Sobre Nosotros", "content": "", "highlights": []},
-                "services": {"title": "Servicios", "subtitle": "", "items": []},
-                "products": {"title": "Productos", "subtitle": "", "items": []},
-                "testimonials": {"title": "Testimonios", "items": []},
-                "gallery": {"title": "Galería", "subtitle": "", "items": []},
-                "pricing": {"title": "Precios", "subtitle": "", "items": []},
-                "faq": {"title": "Preguntas Frecuentes", "items": []},
-                "contact": {"title": "Contacto", "subtitle": ""},
-            }
-            content[section_id] = section_defaults.get(section_id, {"title": "", "content": ""})
-
-        # Guardar variante si viene
-        if variant:
-            content[section_id]["_variant"] = variant
-
-        # Actualizar orden
-        order = content.get("_section_order", [])
-        if order and section_id not in order:
-            order.append(section_id)
-            content["_section_order"] = order
-
-        config.content_data = content
-        config.save(update_fields=["content_data", "updated_at"])
-
-        return Response(
-            {"message": f"Sección '{section_id}' agregada", "section": content[section_id]},
-            status=status.HTTP_201_CREATED,
-        )
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-class RemoveSectionView(APIView):
-    """
-    POST /api/websites/sections/remove/
-
-    Elimina una sección del content_data.
-    Body: {"section_id": "gallery"}
-    No permite eliminar secciones requeridas.
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_404_NOT_FOUND)
-
-        section_id = request.data.get("section_id", "")
-        if not section_id:
-            return Response({"error": "Se requiere 'section_id'"}, status=status.HTTP_400_BAD_REQUEST)
-
-        # Verificar que no sea requerida
-        structure = config.template.structure_schema or {}
-        for s in structure.get("sections", []):
-            if s["id"] == section_id and s.get("required", False):
-                return Response(
-                    {"error": f"La sección '{section_id}' es requerida y no se puede eliminar"},
-                    status=status.HTTP_400_BAD_REQUEST,
-                )
-
-        content = config.content_data or {}
-        if section_id not in content:
-            return Response(
-                {"error": f"La sección '{section_id}' no existe en el contenido"}, status=status.HTTP_404_NOT_FOUND
-            )
-
-        del content[section_id]
-
-        # Actualizar orden
-        order = content.get("_section_order", [])
-        if order and section_id in order:
-            order.remove(section_id)
-            content["_section_order"] = order
-
-        config.content_data = content
-        config.save(update_fields=["content_data", "updated_at"])
-
-        return Response({"message": f"Sección '{section_id}' eliminada"})
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# DUPLICAR SECCIÓN
-# ===================================
-
-
-class DuplicateSectionView(APIView):
-    """
-    POST /api/websites/sections/duplicate/
-
-    Duplica una sección existente con todo su contenido.
-    Body: {"section_id": "services"}
-    Genera un id único: services_2, services_3, etc.
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_404_NOT_FOUND)
-
-        section_id = request.data.get("section_id", "")
-        if not section_id:
-            return Response({"error": "Se requiere 'section_id'"}, status=status.HTTP_400_BAD_REQUEST)
-
-        content = config.content_data or {}
-        if section_id not in content:
-            return Response(
-                {"error": f"La sección '{section_id}' no existe en el contenido"}, status=status.HTTP_404_NOT_FOUND
-            )
-
-        # Generar id único para la copia
-        import copy
-        import re
-
-        # Derive base type: use existing _type, or strip trailing _N suffix
-        section_data = content[section_id]
-        if isinstance(section_data, dict) and "_type" in section_data:
-            base_id = section_data["_type"]
-        else:
-            base_id = re.sub(r"_\d+$", "", section_id)
-        counter = 2
-        new_id = f"{base_id}_{counter}"
-        while new_id in content:
-            counter += 1
-            new_id = f"{base_id}_{counter}"
-
-        # Copiar contenido de la sección original
-        cloned = copy.deepcopy(content[section_id])
-        # Preserve the base section type so render dispatch works for duplicated ids
-        if "_type" not in cloned:
-            cloned["_type"] = base_id
-        content[new_id] = cloned
-
-        # Insertar justo después del original en el orden
-        order = content.get("_section_order", [])
-        if order and section_id in order:
-            idx = order.index(section_id)
-            order.insert(idx + 1, new_id)
-            content["_section_order"] = order
-
-        config.content_data = content
-        config.save(update_fields=["content_data", "updated_at"])
-
-        return Response(
-            {
-                "message": f"Sección '{section_id}' duplicada como '{new_id}'",
-                "new_section_id": new_id,
-                "section": content[new_id],
-            },
-            status=status.HTTP_201_CREATED,
-        )
-
-    def _get_tenant(self, request):
-        if not hasattr(request.user, "tenant") or not request.user.tenant:
-            return None
-        return request.user.tenant
-
-
-# ===================================
-# CAMBIAR VARIANTE DE SECCIÓN
-# ===================================
-
-
-class UpdateSectionVariantView(APIView):
-    """
-    POST /api/websites/sections/variant/
-
-    Cambia la variante visual de una sección.
-    Body: {"section_id": "hero", "variant": "split-image"}
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    VALID_VARIANTS = {
-        "hero": ["centered", "split-image", "fullwidth-image", "bold-typography", "diagonal-split", "glassmorphism"],
-        "about": ["text-only", "split-image", "stats-banner", "timeline", "overlapping-cards", "fullwidth-banner"],
-        "services": [
-            "grid-cards",
-            "grid-cards-image",
-            "list-detailed",
-            "featured-highlight",
-            "horizontal-scroll",
-            "icon-minimal",
-        ],
-        "products": [
-            "grid-cards",
-            "grid-cards-image",
-            "showcase-large",
-            "catalog-compact",
-            "masonry-staggered",
-            "price-table",
-        ],
-    }
-
-    def post(self, request):
-        tenant = self._get_tenant(request)
-        if not tenant:
-            return Response({"error": "Usuario no asociado a un tenant"}, status=status.HTTP_400_BAD_REQUEST)
-
-        try:
-            config = WebsiteConfig.objects.get(tenant=tenant)
-        except WebsiteConfig.DoesNotExist:
-            return Response({"error": "No tienes un sitio web configurado"}, status=status.HTTP_404_NOT_FOUND)
-
-        section_id = request.data.get("section_id", "")
-        variant = request.data.get("variant", "")
-
-        if not section_id or not variant:
-            return Response({"error": "Se requieren 'section_id' y 'variant'"}, status=status.HTTP_400_BAD_REQUEST)
-
-        # Validar variante
-        valid = self.VALID_VARIANTS.get(section_id)
-        if not valid:
-            return Response(
-                {"error": f"La sección '{section_id}' no soporta variantes"}, status=status.HTTP_400_BAD_REQUEST
-            )
-
-        if variant not in valid:
-            return Response(
-                {"error": f"Variante '{variant}' no válida. Opciones: {', '.join(valid)}"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        content = config.content_data or {}
-        if section_id not in content:
-            return Response(
-                {"error": f"La sección '{section_id}' no existe en el contenido"}, status=status.HTTP_404_NOT_FOUND
-            )
-
-        content[section_id]["_variant"] = variant
-        config.content_data = content
-        config.save(update_fields=["content_data", "updated_at"])
-
-        return Response(
-            {
-                "message": f"Variante de '{section_id}' cambiada a '{variant}'",
-                "section_id": section_id,
-                "variant": variant,
-            }
-        )
 
     def _get_tenant(self, request):
         if not hasattr(request.user, "tenant") or not request.user.tenant:
