+++ frontend/.dockerignore
@@ -1,0 +1,6 @@
+node_modules
+.next
+.git
+.env*.local
+npm-debug.log*
+.DS_Store
